refactor: move files to more appropriate packages and do some more specific renaming
